+++ samples/src/main/kotlin/jp/shiguredo/sora/sample/ui/MainActivity.kt
@@ -25,6 +25,7 @@ class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
 
         SoraLogger.enabled = true
+        SoraLogger.libjingle_enabled = false
 
         super.onCreate(savedInstanceState)
         MainActivityUI().setContentView(this)